# api/Identify.go
package api

import (
	"net/http"

	"github.com/labstack/echo"
)

// IdentifyRequest represents the body of an Identify request.
type IdentifyRequest struct {
	UserId       string                 `json:"userId"`
	AnonymousId  string                 `json:"anonymousId"`
	Context      map[string]interface{} `json:"context"`
	Integrations map[string]interface{} `json:"integrations"`
	Timestamp    string                 `json:"timestamp"`
	Traits       map[string]interface{} `json:"traits"`
}

// IdentifyResponse represents the body of an Identify response.
type IdentifyResponse struct {
	// Add any fields you want to return in the response here
}

// IdentifyHandler handles incoming Identify requests.
func IdentifyHandler(ctx echo.Context) error {
	req := IdentifyRequest{}
	if err := ctx.Bind(&req); err != nil {
		return echo.ErrBadRequest
	}

	// Check for required fields (userId or anonymousId)
	if req.UserId == "" && req.AnonymousId == "" {
		return echo.NewHTTPError(http.StatusBadRequest, "Missing required field: userId or anonymousId")
	}

	resp := doSthWithRequest(req)
	return ctx.JSON(http.StatusOK, resp)
}

func doSthWithRequest(req IdentifyRequest) IdentifyResponse {
	return IdentifyResponse{}
}