# swagger.yaml
swagger: '2.0'
info:
  title: RudderStack HTTP API
  version: 1.0.0
  description: >-
    The RudderStack HTTP API lets you track your customer event data and route
    it to your desired destinations.
  termsOfService: 'https://www.rudderstack.com/master-service-agreement/'
  contact:
    email: support@rudderstack.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
basePath: /v1
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
paths:
  /api/v1/identify:
    post:
      tags:
        - HTTP API
      summary: Identify
      description: >-
        The identify call lets you associate a visiting user to their actions
        and record any associated traits.
      operationId: Identify
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/IdentifyPayload'
      responses:
        '200':
          description: OK
...
definitions:
  IdentifyPayload:
    type: object
    properties:
      userId:
        type: string
        description: Unique identifier for a particular user in your database.
      anonymousId:
        type: string
        description: >-
          Sets the user ID for cases where there is no unique identifier for the
          user.
      context:
        type: object
        description: Dictionary of information that provides context about a message.
      timestamp:
        type: string
        format: date-time
        description: The timestamp of the message’s arrival.
...